pyproject.toml
[project]
name = "agentic-systems"
version = "0.1.0"
description = "Exploration of agentic systems using LangChain and LangGraph."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.1",
    "fastapi[standard]>=0.119.1",
    "fastmcp>=2.12.5",
    "grandalf>=0.8",
    "langchain>=1.0.2",
    "langchain-anthropic>=0.3.20",
    "langchain-azure-ai>=1.0.0",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.76",
    "langchain-mcp-adapters>=0.1.11",
    "langchain-nomic>=0.1.5",
    "langchain-ollama>=0.3.8",
    "langchain-qdrant>=0.2.1",
    "langchain-text-splitters>=0.3.11",
    "langchain-xai>=1.0.0",
    "langgraph>=1.0.2",
    "langgraph-api==0.5.0",
    "langgraph-cli[inmem]>=0.3.3",
    "nltk>=3.9.1",
    "nomic[local]>=3.5.3",
    "qdrant-client>=1.15.1",
    "tiktoken>=0.11.0",
    "unstructured[md]>=0.18.14",
]
